% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteredinterference.R
\docType{package}
\name{clusteredinterference}
\alias{clusteredinterference}
\alias{clusteredinterference-package}
\title{Draw Causal Inference from Observational Studies with Clustered Interference}
\description{
\pkg{clusteredinterference} implements the estimators from Barkley et al.,
(201X), \emph{Causal Inference from Observational Studies with Clustered
Interference}. These are inverse probability of treatment-weighted estiamtors
for estimating the causal effects on outcomes of population-level treatment
policies.
}
\details{
These estimators are implemented in one function, \code{\link{policyFX}}. A
healthy amount of documentation can be found in the vignette included in this
package.
}
