% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator.R
\name{policyFX}
\alias{policyFX}
\title{Estimate causal effects (FX) of nonrandomized policies}
\usage{
policyFX(data, formula, alphas, k_samps = NULL, ..., verbose = FALSE,
  root_options = NULL, nAGQ = 2, return_matrices = FALSE)
}
\arguments{
\item{data}{a dataframe (not a tibble)}

\item{formula}{a long formula call with outcome | treatment ~ covariates + (1|cluster_id) | cluster_id}

\item{alphas}{a vector of policies of interest. Each entry must be between 0 and 1}

\item{k_samps}{The maximum number of vectors to evaluate to estimate the omega term. Set to 0 to leave upcapped without a maximum. Recommended to set <= 5.}

\item{...}{The dots}

\item{verbose}{if TRUE then will print (lots of) output. Defaults to FALSE.}

\item{root_options}{passed to multiroot function.}

\item{nAGQ}{passed into lme4::glmer(). Recommended more than 1. Defaults to 2.}

\item{return_matrices}{whether to return variance matrices}
}
\value{
a list
}
\description{
For the dots function, user may supply their own \code{target_grid} through the dots argument. The \code{target_grid} can be made through exported function \code{makeTargetGrid}
}
