% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.policyFX}
\alias{print.policyFX}
\title{Prints a summary of the estimates from a policyFX object}
\usage{
\method{print}{policyFX}(x, ...)
}
\arguments{
\item{x}{object of class "policyFX"}

\item{...}{User may specify integer \code{nrows}.}
}
\description{
Prints a summary of the estimates from a policyFX object
}
