% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{summary.policyFX}
\alias{summary.policyFX}
\title{Prints a summary of a policyFX object}
\usage{
\method{summary}{policyFX}(object, ...)
}
\arguments{
\item{object}{object of class "policyFX"}

\item{...}{User may specify integer \code{nrows}.}
}
\description{
Prints a summary of a policyFX object
}
\author{
Brian G. Barkley
}
