% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{A "toy" dataset for illustrating the estimator}
\format{A data frame with 98 rows and 5 variables: \describe{
  \item{Outcome}{Individual outcome status observed at the end of follow-up.
  E.g., infection status} \item{Treatment}{Individual treatment status
  observed at the end of follow-up. E.g., vaccination status}
  \item{Cluster_ID}{Unique identifier for the cluster of which the individual
  is a member} \item{Age}{Individual's age, in years}
  \item{Distance}{Individual's distance to river, in kilometers} }}
\usage{
toy_data
}
\description{
A dataset containing simulated data for outcome, treatment status, cluster
membership, and two pretreatment covariates. There are 22 clusters that have
3 individuals each, and 8 clusters that have 4 individuals each.
}
\keyword{datasets}
