% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTargetGrid.R
\name{makeTargetGrid}
\alias{makeTargetGrid}
\title{Creates the grid of all estimands to estimate.}
\usage{
makeTargetGrid(alphas, small_grid = FALSE)
}
\arguments{
\item{alphas}{A numeric vector for the probabilities corresponding to the
policies of interest. Each entry must be between 0 and 1.}

\item{small_grid}{if TRUE then only estimates some policies. Default is FALSE.}
}
\description{
Creates the grid of all estimands to estimate.
}
